Add keyring signer and unlock fail feedback (#176)
