+++ app/App/Panel/Main/Add/style/index.styl
@@ -336,7 +336,7 @@
 
       .addAccountItemDeviceTitle
         font-weight 600
-        padding-bottom 4px
+        text-align center
 
       .addAccountItemDeviceStatus
         font-size 10px
@@ -391,7 +391,7 @@
 
       .addAccountItemOptionSetupFrames
         display flex
-        width 400%
+        width 1000%
         height 100%
         background rgba(5, 25, 35, 0.2)
 
@@ -401,7 +401,7 @@
           justify-content center
           flex-direction column
           transition standardSlow
-          width 25%
+          width 10%
 
           .addAccountItemOptionTitle
             display flex
@@ -421,16 +421,6 @@
 
     .addAccountItemOptionSubmit:hover
       background rgba(5, 25, 35, 0.3)
-
-    // .addAccountItemOptionInput:before
-    //   content ''
-    //   position absolute
-    //   bottom 0px
-    //   left 20px
-    //   right 20px
-    //   height 4px
-    //   border-radius 2px
-    //   background rgba(5, 25, 35, 0.5)
 
     .addAccountItemOptionList
       width calc(100% - 20px)
