+++ app/App/Panel/Main/Signer/index.js
@@ -115,15 +115,16 @@ class Signer extends React.Component {
                 if (this.props.signer) {
                   if (this.props.signer.type === 'ledger') return <img src={ledgerLogo} />
                   if (this.props.signer.type === 'trezor') return <img className='trezorImage' src={trezorLogo} />
-                  if (this.props.signer.type === 'hot') return svg.flame(21)
-                  if (this.props.signer.type === 'ring') return svg.octicon('ring', { height: 31 })
+                  if (this.props.signer.type === 'seed' || this.props.signer.type === 'ring') return svg.flame(21)
                   return svg.octicon('plus', { height: 31 })
                 } else {
                   return null
                 }
               })()}
             </div>
-            <div className='signerText'>{this.props.signer ? this.props.signer.type : 'no signer'}</div>
+            <div className='signerText'>{this.props.signer ? (
+              this.props.signer.type === 'ring' || this.props.signer.type === 'seed' ? 'hot' : this.props.signer.type
+            ) : 'no signer'}</div>
           </div>
         </div>
         <div className='addressSelect' style={this.store('selected.view') === 'settings' ? { opacity: 1, transitionDelay: '0s' } : { opacity: 0, transitionDelay: '0.2s' }} onMouseDown={e => {
